\name{dental}
\alias{dental}
\title{Measurements on 11 girls and 16 boys, at 4 different ages}
\description{Potthoff and Roy (1964) introduce this data set. Measurment is the distance, in millimeters, from the center of pituitary to the pterygomaxillary fissure of 11 girls and 16 boys at ages 8, 10, 12 and 14.}

\source{
  Potthoff, R.F. and Roy, S.N. (1964). 
  A Generalized Multivariate Analysis of Variance Model Useful Especially for Growth Curve Problems.
  \emph{Biometrika} Vol. 51, No. 3/4 (Dec., 1964), 313-326.
}

\docType{data}

